Added spPostInvTransaction with spPostInvTransDetail

+++ IspDataDb/dbo/Tables/SaleDetail.sql
@@ -3,15 +3,15 @@
     [Sequence]       SMALLINT        NULL,
     [SaleIdNo]       INT             NULL,
     [ProductIdNo]    INT             NULL,
-    [Quantity]       SMALLINT        NULL,
+    [Quantity]       DECIMAL(12, 4)        NULL,
     [UnitIdNo]       TINYINT         NULL,
     [BatchNo]        VARCHAR (10)    NULL,
     [Price]          DECIMAL (9, 2)  NULL,
-    [DiscountAmount] DECIMAL (12, 2) NULL,
+    [DiscountAmount] DECIMAL (9, 2) NULL,
     [ExpiryDate]     DATE            NULL,
-    [VatAmount]      DECIMAL (12, 2) NULL,
+    [VatAmount]      DECIMAL (9, 2) NULL,
     [VatPercent]     DECIMAL (5, 2)  NULL,
-    [NetAmount]      DECIMAL (12, 2) NULL,
+    [NetAmount]      DECIMAL (9, 2) NULL,
     CONSTRAINT [PK_SaleDetail] PRIMARY KEY CLUSTERED ([IdNo] ASC)
 );
